pyproject.toml
[tool.poetry]
name = "aiagent-multiple-database-chat"
version = "0.1.0"
description = ""
authors = ["supakorn.emch <supakorn.emch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
